Refactor: Move RoadRunner resources to dedicated package
Moved RoadRunner-specific classes (Localizer, MecanumDrive, TankDrive, etc.) from the main `org.firstinspires.ftc.teamcode` package to a new `RoadRunnerResourses` package.

Updated import statements in tuning OpModes to reflect the new package structure.

Removed an unused import (`DcMotorSimple`) from `MecanumDrive.java` and `TankDrive.java`.

Add: Basic TeleOp for variable practice

Created a new OpMode `VariablePractice.java` that initializes and displays several variable types (int, double, boolean, String) using telemetry. This OpMode has an empty loop method.
